--- src/components/CommentList.js
@@ -1,57 +1,0 @@
-import React, {Component} from 'react';
-import Comment from './Comment';
-import PropTypes from 'prop-types';
-import toggleOpen from '../decorators/toggleOpen';
-
-class CommentList extends Component {
-    componentDidMount() {
-        console.log('---', 'mounted')
-    }
-
-    componentWillUnmount() {
-        console.log('---', 'unmounting')
-    }
-
-    componentWillReceiveProps() {
-        console.log('---', 'updating')
-    }
-
-    render() {
-        const {comments, isOpen, toggleOpen} = this.props;
-        const text = isOpen ? 'hide comments' : 'show comments';
-        return (
-            <div>
-                <button onClick={toggleOpen}>{text}</button>
-                {getBody(this.props)}
-            </div>
-        );
-    }
-}
-
-function getBody(props) {
-    const {comments, isOpen} = props;
-    if (!isOpen) return null;
-
-    const body = comments.length ? (
-        <ul>
-            {comments.map(comment => <li key={comment.id}><Comment comment={comment}/></li>)}
-        </ul>
-    ) : <h3>No comments yet</h3>;
-
-    return (
-        <div>
-            {body}
-        </div>
-    );
-}
-
-
-CommentList.defaultProps = {
-    comments: []
-};
-
-CommentList.propTypes = {
-    comments: PropTypes.array.isRequired
-};
-
-export default toggleOpen(CommentList);
