+++ src/components/CommentList/style.css
@@ -1,0 +1,4 @@
+#comment {
+	width: 400px;
+	height: 200px;
+}
