+++ src/decorators/toggleOpen.js
@@ -2,7 +2,8 @@ import React, {Component as ReactComponent} from 'react';
 
 export default (OriginalComponent) => class DecoratedComponent extends ReactComponent {
     state = {
-        isOpen: false
+        isOpen: false,
+		commentText: ''
     };
 
     toggleOpen = () => this.setState({
